+++ apps/router-e2e/__e2e__/03-server-actions-only/app/index.tsx
@@ -1,0 +1,13 @@
+import 'client-only';
+import React from 'react';
+import { Text } from 'react-native';
+
+import { renderPage } from '../components/server-actions';
+
+export default function ServerActionTest() {
+  return (
+    <React.Suspense fallback={<Text>Loading...</Text>}>
+      {renderPage({ title: 'Hello!' })}
+    </React.Suspense>
+  );
+}